refactor: clean up src/phoenix/trace/v1
